Move http-transport module into superagent directory
